3.1.7 Rename package
