+++ src/test/java/edu/kis/vh/nursery/RhymersJUnitTest.java
@@ -3,7 +3,7 @@ package edu.kis.vh.nursery;
 import org.junit.Assert;
 import org.junit.Test;
 
-import edu.kis.vh.nursery.list.IntArrayStack;
+import edu.kis.vh.nursery.collections.IntArrayStack;
 
 public class RhymersJUnitTest {
